+++ resources/views/combobox.blade.php
@@ -1,13 +1,16 @@
-<div class="{{ $comboboxContainerClass ?? 'flex' }}">
+<div class="{{ $config['containerClass'] ?? 'flex' }}">
 
     {{-- Elements --}}
     @foreach ($elements as $element)
         {{-- Include a select field --}}
-        @includeWhen(Combobox::value($element, 'type') === 'select', 'livewire-combobox::type.select')
+        @includeWhen(
+            $element->getType() === 'select',
+            'livewire-combobox::type.select'
+        )
     @endforeach
 
     {{-- Loading --}}
-    @if ($loading)
+    @if ($config['loading'])
         @include('livewire-combobox::loading')
     @endif
 </div>