+++ resources/views/type/select.blade.php
@@ -1,54 +1,50 @@
 @php
-    $totalOptions = Combobox::count($element, 'options');
+    $totalOptions = count($element->options);
 @endphp
 
-@unless (Combobox::value($element, 'hideOnEmpty') === true && $totalOptions === 0)
+@unless ($totalOptions <= 0 && $element->getHideOnEmpty() === true)
     {{-- Field main container --}}
     <div
-        id="field-container-for-{{ Combobox::value($element, 'uriKey') }}"
-        class="{{ Combobox::value($element, 'fieldContainerCss') ?? 'px-2' }}"
+        id="field-container-for-{{ $element->getUriKey() }}"
+        class="{{ $element->getFieldContainerCss() ?? 'px-2' }}"
     >
         {{-- Label --}}
         <label
-            id="label-for-{{ Combobox::value($element, 'uriKey') }}"
-            class="{{ Combobox::value($element, 'labelCss') ?? 'block w-full text-sm font-medium text-gray-600' }}"
-            for="{{ Combobox::value($element, 'uriKey') }}"
+            id="label-for-{{ $element->getUriKey() }}"
+            class="{{ $element->getFieldLabelCss() ?? 'block w-full text-sm font-medium text-gray-600' }}"
+            for="{{ $element->getUriKey() }}"
         >
-            {{ Combobox::value($element, 'label') }}
+            {{ $element->label }}
         </label>
         {{-- Field --}}
         <select
-            id="{{ Combobox::value($element, 'uriKey') }}"
-            dusk="{{ Str::of(Combobox::value($element, 'uriKey'))->replace('-', '_') }}"
-            name="{{ $element?->label  ?? $element['uriKey'] }}"
-            class="{{ Combobox::value($element, 'fieldCss') ?? 'bg-white w-full relative border border-gray-300 rounded-md shadow-sm pl-3 pr-10 py-2 mt-1 text-left cursor-default focus:outline-none focus:ring-1 focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm' }}"
+            id="{{ $element->getUriKey() }}"
+            dusk="{{ Str::of($element->getUriKey())->replace('-', '_') }}"
+            name="{{ $element?->label  ?? $element->getUriKey() }}"
+            class="{{ $element->getFieldCss() ?? 'bg-white w-full relative border border-gray-300 rounded-md shadow-sm pl-3 pr-10 py-2 mt-1 text-left cursor-default focus:outline-none focus:ring-1 focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm' }}"
 
             {{-- Disable empty elements --}}
-            @if($totalOptions <= 0)
+            @if($element->disabledOnEmpty() && $element->getTotalOptions() <= 0)
                 disabled
             @endif
 
             {{-- Last element or element without Livewire Response --}}
-            @unless (Combobox::value($element, 'withoutResponse'))
-                wire:model.defer="comboboxValues.{{ Combobox::value($element, 'uriKey') }}"
-                wire:change="$emit('dependOn', '{{ Combobox::value($element, 'uriKey') }}')"
+            @unless ($element->getWithoutResponse())
+                wire:model.defer="comboboxValues.{{ $element->getUriKey() }}"
+                wire:change="$emit('dependOn', '{{ $element->getUriKey() }}')"
             @endunless
         >
             {{-- Show or hide empty first element --}}
-            @unless (Combobox::value($element, 'firstRemoved'))
+            @unless ($element->getFirstRemoved())
                 <option value=""></option>
             @endunless
 
             {{-- Options --}}
-            @foreach (Combobox::value($element, 'options') as $key => $value)
+            @foreach ($element->getOptions() as $key => $value)
                 <option
-                    value="{{ $key ?? Combobox::value($element, 'defaultValue') }}"
-                    {{-- Selected option for parents --}}
-    {{--                 @if(is_object($element) && $element->defaultValue == $key)
-                        selected
-                    @endif --}}
+                    value="{{ $key ?? $element->defaultValue }}"
                     {{-- Selected option for childs --}}
-                    @if(is_object($element) && $element->defaultValue === $key)
+                    @if(is_object($element) && $element->getDefaultValue() === $key)
                         selected
                     @endif
                 >