+++ src/LivewireComboboxServiceProvider.php
@@ -4,9 +4,6 @@ declare(strict_types=1);
 
 namespace Daguilarm\LivewireCombobox;
 
-use Daguilarm\LivewireCombobox\Facades\Combobox;
-use Daguilarm\LivewireCombobox\Facades\ComboboxProvider;
-use Illuminate\Foundation\AliasLoader;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 
@@ -17,11 +14,5 @@ final class LivewireComboboxServiceProvider extends PackageServiceProvider
         $package
             ->name('livewire-combobox')
             ->hasViews();
-    }
-
-    public function registeringPackage(): void
-    {
-        $this->app->register(ComboboxProvider::class);
-        AliasLoader::getInstance()->alias('Combobox', Combobox::class);
     }
 }