+++ tests/App/Http/Livewire/ComboboxSelects.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace Daguilarm\LivewireCombobox\Tests\App\Http\Livewire;
 
-use Daguilarm\LivewireCombobox\Components\ComboboxLivewireComponent;
+use Daguilarm\LivewireCombobox\Components\ComboboxComponent;
 use Daguilarm\LivewireCombobox\Components\Fields\Select;
 use Daguilarm\LivewireCombobox\Contracts\Combobox;
 use Daguilarm\LivewireCombobox\Tests\App\Models\Car;
 use Daguilarm\LivewireCombobox\Tests\App\Models\Extra;
 use Daguilarm\LivewireCombobox\Tests\App\Models\Option;
 
-class ComboboxSelects extends ComboboxLivewireComponent implements Combobox
+class ComboboxSelects extends ComboboxComponent implements Combobox
 {
     public function elements(): array
     {