+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `Livewire Combobox` will be documented in this file:
 
+## 1.2 - 2021-05-11
+
+- New code structure. All the code has been refactoring and optimized.
+- Major security improvements, and now, the properties are `protected` not `public` as before.
+- New method `disabledOnEmpty()`.
+- The variable `public string $comboboxContainerClass` now is `protected string $containerClass`.
+
 ## 1.1 - 2021-05-04
 
 - New loading indicator.